move gitattributes to root
